Optimize day10
Benchmarking shows that the majority of the time is spend in extract_path.
I used `Vec::contains` in my initial implementation and never changed it. That's
lurking quadratic behavior.

I considered changing it to a HashSet, but given that we're looking for a totally closed loop,
we're guaranteed that the loop will end where it started.

The benchmarks show a 96% improvement
```
change: [-96.466% -96.448% -96.430%] (p = 0.00 < 0.05)
```